requirements/development.txt
-r base.txt

django-debug-toolbar>=4.0.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0